requirements.txt
Django
djangorestframework
gunicorn
dj-database-url
psycopg2
matplotlib
Pillow
whitenoise